add option to include subtrees in full_tree
git allows for a '-t' flag which will include subtrees when getting the
full contents of a tree. this a nice option to have when calling
full_tree, so rather than creating a separate method I've added a opts
parameter that translates a :subtrees option to the corresponding flag
to the git command.